(#8) Move native events/handlers to new partial
